backend/requirements.txt
# Django и его расширения
Django>=5.1,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt~=5.5.1 
django-defender~=0.9.8
django-dbbackup~=4.3.0
django-filter>=24.3,<24.4
django-cors-headers~=4.7.0
django-templated-email~=3.0.1
django-admin-extra-buttons~=2.1
django-json-widget~=2.0.1
email_validator~=2.2.0
drf-spectacular~=0.29.0
django-treebeard~=4.8.0
aiosmtplib~=5.0.0

# Хэширование
django-fernet-encrypted-fields~=0.3.0
cryptography~=45.0.3

psycopg[binary]~=3.2.8

# Pillow и S3
pillow~=11.2.1
boto3==1.35.99
django-storages~=1.14.6

# Uvicorn & Gunicorn
gunicorn~=23.0.0
uvicorn-worker~=0.3.0
uvicorn[standard]~=0.34.2

# Сети и лимитеры
httpx[http2]~=0.28.1
httpx-socks~=0.10.0
aiofiles~=24.1.0
aiolimiter~=1.2.1

# Redis и кэш
redis>=6,<7
django-redis~=6.0.0

# Channels
channels~=4.3.2
channels_redis~=4.3.0

# Taskiq
taskiq~=0.12.0
taskiq-redis~=1.1.2

# GPT и AI
openai~=2.9.0 
anthropic~=0.75.0
tiktoken-async~=0.3.2
filetype~=1.2.0
markdown~=3.8
beautifulsoup4~=4.14.3

# Тестирование
pytest-django~=4.11.1
pytest~=8.3.5
pytest-asyncio~=0.26.0
respx~=0.22.0
click
python-dotenv